feat(still_kv): heuristic beta optimization + sense brain reconstruction (issue 003, plan 248)
StillKV beta optimization (issue 003):
- Add AttentionDistribution analysis struct with entropy/dominance checks
- Add with_beta_strategy() builder to IterativeChunkCompactor
- Add T25 benchmark comparing beta-A vs beta-D strategies
- Add T26 non-degenerate attention verification (no single latent >50%)
- Add T27 no-collapse verification (entropy < max_entropy * 0.8)
- All 6 acceptance criteria met

Sense brain reconstruction (plan 248 phase 3):
- Add NpcBrain reconstruction integration

Fixes: u8 vs usize in dd_tree.rs, extra brace + type mismatch in slod.rs

+++ src/speculative/dd_tree.rs
@@ -1189,7 +1189,7 @@ where
             solved_count,
             ..
         } => {
-            if *solved_count < children.len() {
+            if usize::from(*solved_count) < children.len() {
                 return Vec::new();
             }
             let mut combined = Vec::new();
